[FIX] website: fixed issue with redirect old url list
when user click on 'could be used in' a list opens up
the list was not rendering properly i.e. being cut with no scroll previeusly
also, the list was appearing behind modal when viewed with a small screen device

we changed postion of the menu outside dialog box to view it clearly and
provide scrolling for long list. Also when on small screens, menu was rendering
behind the modal or popup, not able to see

task-2092269
Closes https://github.com/odoo/odoo/pull/40918