[IMP] web_editor: introduce many2one userValueWidget for reuse
Previously, edition of many2ones in the front-end was done through a
snippet option. While this works fine for raw many2ones in the page, it
is desirable that other snippet options can leverage a many2one as a
part of a more complete or comprehensive option. This commit introduces
a many2one userValueWidget that will be able to be reused by snippet
options, and adapts the existing many2one snippet option to use it
instead.

The interface has also been reworked so as to lean on the existing
we-select userValueWidget for the sake of interface consistency.

task-2190603