[IMP] account: notify sales rep when their invoices are validated
Purpose of this task is to allow sales rep(who don't have any billing right) to
be aware when invoices related to their SO have been approved by accounting so
that they can send them to their customer.

So in this commit, 'validated' is subscribed by default on the user so whenever
one of their colleagues approved created invoice then they are notified.

Task-ID: 2418033