[FIX] point_of_sale: fix thai character incorrectly displayed on email
Some thai character do display incorrectly in the email sent from the PoS.
It can be reproduced by using a name such as: "สมภาร สุมิตร (ร้านภารทองปลื้ม)". Note that the issue only appear in the email being sent to the customer, not the receipt preview shown in Odoo PoS.

This is due to a bug in the html2canvas library, see: https://github.com/niklasvh/html2canvas/issues/740
A fix was proposed on the library in 2016 but wasn't merged:
https://github.com/niklasvh/html2canvas/pull/778
I did tested it locally and it looks to work like a charm.

OPW-2447562