[FIX] mrp: fix rounding issue for negative kit quantity
For kit with the negative quantity, we always round down the value
(example: If I have only -0.2 component, we consider that we have -1
kit).
Also the fields compute by `_compute_quantities_dict` doesn't have
the same rounding than quant (rounded depending of the product UoM vs
Product Unit of Measure (which can be more precise)). Then we can have
different result during and after the `_compute_quantities_dict`
and due of this https://github.com/odoo/odoo/pull/65701 the compute of
kit quantity can be wrong depending the which record is in the prefetch.
Example:
- Set default decimal precision to 5 and the rounding precision of Unit
is 4.
- Have a product A and B (both Unit). A is a kit with B as the only
component.
- You have a quantity -384 of B.
- Compute the `qty_available` only for A (it should be -384):
    - The `_compute_quantities_dict` is called for A
    - The `_compute_quantities_dict` of A need the `qty_available`
      of B (for kit computation quantity)
    - The `_compute_quantities_dict` is called for B and will return
      -384.0 but then it will rounded with 5 decimal precision (because
      of the `qty_available` fields digits). And B.qty_available will
      return float_round(-384.0, 5) = -384.00000000000006.
    - and `-384.00000000000006 // 1` will return `-385` and then
      the remaining computation of A will return `-385`.

upg-6588