[FIX] mail: Message from mailing channel should not make a notification in Odoo
currently,
Message from mailing channel make a notification in Odoo for users
with notification "Handled by Email" when user is not log-in. this is happening
because mailing channel is marked ans seen from
"messaging_notification_handler.js" which  is not executed when user is not
log-in and receive message so message counter is increase when user log-in.

after this commit,
Message from mailing channel does not make a notification in Odoo for users with
notification "Handled by Email".to achieve that mark mailing channel as seen if
user has notification "Handled by Email" from "thread.js" when  unread-massage
counter is calculated.

ID - 2541660