[FIX] mail: regarding scheduling activity in list view
Open a activity view(record modal), that will show records, clicking on the
'o_activity_btn' of a record will show dropdown and make sure that there are
many scheduled activity there .

Before this commit:

The dropdown was bigger than the record modal which had a scroll bar. It was not
scrollable properly and "+ Schedule Activity" button is not accessible.

After this commit:

The dropdown follows the scroll, and it will be possible to click the button.

Task- 2528111