[IMP] mail, mass_mailing: remove the <mail.mail> in batch
Purpose
=======
Improve the performance of mass mailing by removing the <mail.mail> in
batch. The unlink operation is very expensive for the ORM, in term of
SQL queries, so we removed them in batch.

The main difference is that now the unlink operation is not committed
for each individual <mail.mail>.

Task-2587345