[IMP] survey: add image zoom widget and name
Before this commit, images displayed on survey question answers were small
and hard to read, and nothing existed in order to zoom on them properly.

Also, in backend, only the image was present in columns of question answer
list view. This was not helpful to match a line to an image, as default list
view for binary files shows but the  size of the image.

AFTER THIS COMMIT

When clicking on such an image, the image is zoomed on and displayed in a
foreground modal. The image ratio is kept and size maximized. Also, the images
take most of the available space inside the answer boxes on question display,
making them more readable already.

Also, on backend, add a column in tree view to keep the image name after
loading the image file. This is editable and can be used as an additional
information on image in backend (version, ...)

--- Links ---
Task Id - 2566584
COM PR - odoo/odoo#73718