[IMP] website_slides: fix the broken documentation design
Currently, in the documentation type of courses, card layout for
the lessons is broken because there isn't enough space to display
everything.

To fix this, this commit moves the 'Completed' badge (which is in
the card-footer) and other tags at the end of the card-body, and
it formats the likes and dislikes (ex, 10000 will be displayed
as 10k).

But, that improvement leads to another trouble. Say there are
'1k' likes, and user upvotes the lesson, there is no indication
whether or not the vote was recorded. To fix this, we will now
different icons based on the user vote.

- The thumbs-up icon will be filled if user has liked the lesson,
  and it will be hollow otherwise
- Likewise, the thumbs-down icon will be filled if user has
  disliked the lesson, and it will be hollow otherwise

Clicking on either of the button, if the vote is recorded successfully,
icon and user vote will be toggled for better indication. Note that
before this commit, if user was performing the same action (for example
upvoting again), warning popover used to appear, which isn't the case
anymore because now the vote will simply be toggled. The mechanism is
somewhat similar to youtube (you can also directly dislike the already
liked content and visa-versa).

The same mechanism is also applied for likes/dislikes in the footer
of individual lesson's page. And, on the same page under 'Statistics'
tab, they are simply formatted like everywhere else.

Apart from that, this commit makes other small UI changes listed
as below:

 - We now display three slides per category instead of four, giving
   the cards a bit more space
 - On the lessons' page, we move the likes/dislikes right after the
   description which previously was at bottom right of the page

Upgrade: https://github.com/odoo/upgrade/pull/2854

Task-2607416