[IMP] mail: use email_formatted on partner
Purpose of this commit is to get rid of custom formataddr and instead use
``email_formatted`` field. It holds a more defensive approach of email
address formatting and should lessen issues when dealing with unexpected
input.

Is is now used in

  * mail.mail sending method when sending directly to partners using
    recipient_ids field;
  * mail.channel manual email_to crafting when sending notification emails
    to channel members;
  * computation of email_from on mail.message and various post-alike
    methods of mail.thread;

Task-2612945
COM PR odoo/odoo# 74474
ENT PR odoo/enterprise# 1998