[IMP] portal: avoid error modal on wizard modal superposition
BEFORE THIS COMMIT :

When using the action "grant portal access" on contact, it opens a wizard.
When trying to granting access, an error is raised if email is not properly
formatted or already used by another user, in a modal. The modal of error
message is superimposed on the one of the wizard. This looks very unpleasant
and not very readable.

AFTER THIS COMMIT :

Therefore, the error messages due to an invalid email (bad format or already
used) are simplified and displayed inline in portal wizard user tree view
instead in the form of colored fa-icons, with details as title.
Also, there will be no display of reinvite / grant access buttons as long as
an error on email remains.
To do so, a new selection field sets the email validity status for each
portal_wizard_user : 'valid' / 'invalid' / 'already_registered.
Errors are still thrown if method is called directly, except for the revoke
access method. Indeed, we want to be able to revoke access, independantly of
the email status. However, the partner's email is only updated if the email
is valid.

TESTS :

Tests are updated accordingly. The last assert of test_portal_wizard_error
is repaired. The previous one raised an error as expected but not the good
one. As the email had a wrong format, it was detected and error was raised,
and not because the user was internal, as supposed to. The portal user is
updated as it should be.

--- LINKS ---
Task Id - 2613780
COM PR - odoo/odoo#74911