[FIX] web: keep focus on widget with resetOnAnyFieldChange on navigation
Before this commit: when navigating using keyboard to widget which has
resetOnAnyFieldChange is true, it does not work because resetOnAnyFieldChange
resets widget and renderEdit is called again, on some widget  is replaced
and because of that focusout is called, so focus is not removed from that
widget, expected behaviour is even after reset widget should have focus.

After this commit: when widget is reset and element of widget is replaced
still focus will remain on widget as we called focusLastActivatedWidget method
once reset is done.

task-2636133