[IMP] web_editor: improve performance of collaborative security check
Before this commit, a set of security check were made everytime a
longpolling call was made and for every channels. This set of check were
also made everytime a message was broadcasted to a channel.
In order to avoid checking all the time, a channel signature is
generated once when initializing the editor and is refreshed every hour.

task-2639143