[FIX] website: correctly toggle the preview for the countdown snippet
The countdown widget has an end action that can be configured to show a
message when the countdown reaches zero. A button in the editor toggles
a preview of this message. However, a bug currently makes the preview
disappear whenever the widget is restarted. Such a restart is triggered
whenever the editor emits a widgets_start_request event is emitted in
the editor.

To solve the problem, the preview visibility is now controlled by a
separate css class s_countdown_enable_preview overriding d-none. This
way, the preview visibility no longer interacts with the widget's logic
and is no longer affected by the widget restarting.

task-2638366