[IMP] web_editor: make editor able to apply styles over gradient presets
Before this commit heading gradient presets could not be overridden by
solid colors on a selection in the headers.

After this commit heading gradient presets can be overridden by solid
colors and solid or gradient highlight colors.
A 'text-solid' class is added to all non-'text-gradient' colored texts
to neutralize the effects of a nesting 'text-gradient' or similar, such
as a preset for foreground text gradient color.

task-2633169