[IMP] web_editor: display fully rendered snippet over dropzones
Before this commit, some snippets did not have any or complete preview
when dragging them over a dropzone.

For s_countdown, s_chart, o_facebook_page:
The widget associated to the snippet is started when dragging it.

For s_popup, s_popup_newsletter:
A dropzone taking over all the #wrap is displayed and the modal is shown
when hovering over it.

For s_website_form:
Some elements with class s_preview are added to the template. These
elements are removed when the snippet is dropped.

Also, the class o_colored_level was added to s_three_columns so that the
previews matches the snippet dropped.

task-2431469