[IMP] payment_test:
task-2512196

[IMP] payment_test: split auth&capture + add tests

Add the possibility to test the manual capture with payment test.
Add tests about authorization and capture, and tokenization.

task-2512196

[IMP] payment_test: support cust. fees computation

Add the support of customer fees computation and a test to check it.

task-2512196

[IMP] payment_test: support partial refund

Add the support of partial refund and a test to check it.

task-2512196

[IMP] payment_test: select status of payment

Users will now be able to test pending or canceled transactions directly
through the portal.

task-2512196

[IMP] payment_test: improve pending tx handling

User will be able to change a status from pending to done, authorized
or cancel.

task-2512196