[FIX] stock: no longer update negative move during merge
When trying to absorb quantity from negative move, the product_uom_qty
of the negative move would be updated several times (n-1 times if n
positives moves are required to fully absorb the quantity).

All these updates would trigger the write() method, which would then log
a message in the picking because the move quantity would be updated.
This is valid for positive moves as their quantity is trully updated, but
for the negative it is a temporary state, since it would be unlinked
afterwards.

Task-2659807