[IMP] sale_purchase: Add notification in SO if MTO PO qty is decreased
Log an activity in a Sale Order that warns the user if a quantity to
order is decreased in a related Purchase Order.
This means a sell user could know easily if a Purchase Order necessary
for his Sale Order has been decreased and that it might now not be
enough to fulfill his Sale Order.

Task-2659807