[FIX] website_sale: make cta button work on debug dynamic snippet
Previously, on the debug dynamic snippet, you could choose a product
filter and a template that had an add2cart button but that button
would not work.

This commit fixes that by patching the widget classe instead of
extending it, making the add2cart methods available to any template.

task-2654924