[FIX] web_editor, website: span gradient across text style changes
Since [1] the gradient text image was reset within each font or span
block (e.g. for applying bold style).
The change was done because Chrome does not inherit the involved styles
when a transform is used, which is the case in animated texts.

After this commit the inherit style is only applied on animated texts.
To avoid making the image inherited on * (which causes the gradient
being drawn locally in spans), the background image of the gradient is
applied to the animated text element explicitly in javascript.

[1] https://github.com/odoo/odoo/commit/187acb938f70a2130d25fa76079221339c742f1e

task-2666200