[IMP] portal: allow portal users to deactivate their accounts
Purpose
=======
Allow portal users to deactivate their accounts from the portal view.

After the deactivation, they are redirected to the login page, so
they can verify that they can not longer login with their credentials.

We first try to unlink the user and the partner. If it failed, we
archived them instead and overwrite sensitive information.

After the deactivation, we blacklist the email and the phone of the
user, so we are sure that we never send him again email / SMS.

Task-2629544