[REM] configurator, themes: safari svg previews readability
In Safari (14+) the CSS workaround to make block "readables" over any bg
is broken. By replacing the current CSS workaround by an appropriate
SVG filter in each themes svg, the problem will be fixed.
This commit will remove the CSS workaround.

This change is related to an update of all themes on PR#522 in the
design-themes repo.

task-2667028