[IMP] hr_recruitment: add Interviewer access rights
Create a new Interviewer access rights.

This right is automatically given to users that are added as
Interviewers on a Job Position or on an Application.

An interviewer can only access applications they are interviewer for,
and the chatter is disabled for them as it might contain some sensitive
information (salary details, contract link, etc.)

TaskID: 2669730