[IMP] website: allow mega menu transparency
This PR allows the user to put a transparency effect on the mega menu.
The transparency effect did not work on the mega menu because
the container that contains the mega menu did not have transparency.
Note that the mega menu templates use a default background color
since [1] to not be transparent by default because of this commit.

[1]: https://github.com/odoo/odoo/commit/967d21a8d0dd8c79be1fe0a2d2ec3a6edbb389ca

task-2623335