[FIX] stock: orderpoint doubled qty to order
How to reproduce:
  - Creates a product and creates an orderpoint for this product with
    the following value:
    - min & max at 0;
    - trigger manually;
    - multiple at 1;
  - Creates a delivery for this product and confirms it;
  - Open the Replenishment
    => You should see the orderpoint for this product with the qty. to
       order equals to the qty. to delivery.
  - Set the qty. to order to 0 and re-open the Replenishment
    => The qty. to order is now the double of the qty. to delivered.

The quantity to delivered should be equals to the forecast demand.
The bug was due to the write on the field `qty_forecast` in the method
`_get_orderpoint_action`. As it's a compute field, its value is computed
before to increment its value, so the `qty_forecast` will be equals to
the demand qty. then we add to it the product qty. and so, the forecast
quantity is wrong when we compute the `qty_to_order`.

task-2674053