[FIX] website: prevent deletion of submit buttons and prevent URL edit
Before this commit the user could remove the submit button of forms
and edit the url of other buttons and links that were used as a
ui elements for widgets rather than real links. The edition of buttons
introduced in [1] was also broken.

This commit fixes the deletion of the submit button and edition of the
url. It also re-introduces the edition of button styles.

[1] : fc83af4b0b209ffc0ae3d117fd9865b68aa2f118

task-2669592