[IMP] packaging: update requirement checking tool
In 43dd87851f2, the `requirements-check.py` tool was added in order to
compare requirements.txt with the Debian and Ubuntu distribution.

With this commit, this tool now use subcommands in order to keep the
original goal via the subcommand `distro` and the same arguments as
before.

The second subcommand is `local` and is intended to compare the
`requirements.txt` with the locally installed python packages to find
possible inconsistencies.