[FIX] mail: allow the admin to open any emails
Bug
===
When the admin open an email, he might get an error message if he can
not access one of the attachment of the email.

To fix that, we compute 2 new fields, the attachments to show in the
view (the record the admin have access to), and the number of
attachments we can not read (to display a warning message).

Task-2464212