[IMP] web_editor: add support for fixed size svgs for image shapes
With commit [1] introducing device shapes, it was needed to have SVGs
with a fixed size. Originally the code was created with the idea
that svg will take the size of the image they're going to clip.
This commit introduces a data attribute that allows SVGs to bypass
being resized so that they always keep the same aspect ratio.

[1] : 85768b60c82edd3b45e70852131e3218d75f660b

task-2666049