[IMP] account_payment: hide pay button for auth. tx
Hide the 'Pay now' button for invoices with an authorized transaction.

When a transaction is in the 'authorized' state, the residual amount to
invoice is not updated because the payment is not yet processed, and the
button is not hidden. If the partner made a second payment through the
portal, the user will receive an error when he'll try to capture the
transaction because we can't capture multiple transactions at the same
time.

task-2527323